Refactor: extract ActivitiesView to separate file #3081
